Make localization independent of Flutter
Fixes #935
