+++ flutter/lib/view_models/base/screen_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:delern_flutter/flutter/localization.dart';
 import 'package:delern_flutter/flutter/user_messages.dart';
+import 'package:delern_flutter/l10n/app_localizations.dart';
 import 'package:meta/meta.dart';
 
 abstract class ScreenBloc {
