+++ flutter/lib/views/helpers/send_invite.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import 'package:delern_flutter/flutter/localization.dart';
+import 'package:delern_flutter/flutter/localization.dart' as localizations;
 import 'package:flutter/material.dart';
 import 'package:share/share.dart';
 
 Future<void> sendInvite(BuildContext context) =>
-    Share.share(AppLocalizations.of(context).inviteToAppMessage);
+    Share.share(localizations.of(context).inviteToAppMessage);
